+++ src/octoprint/__main__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python2
+# coding=utf-8
+from __future__ import absolute_import, division, print_function
+
+if __name__ == "__main__":
+	import octoprint
+	octoprint.main()
