+++ src/octoprint/util/gcodeInterpreter.py
@@ -512,85 +512,3 @@ def getCodeFloat(line, code):
 	except:
 		return None
 
-if __name__ == "__main__":
-	import click
-	import time
-	import yaml
-	import sys
-
-	@click.command()
-	@click.option("--throttle", "throttle", type=float, default=None)
-	@click.option("--throttle-lines", "throttle_lines", type=int, default=None)
-	@click.option("--speed-x", "speedx", type=float, default=6000)
-	@click.option("--speed-y", "speedy", type=float, default=6000)
-	@click.option("--speed-z", "speedz", type=float, default=300)
-	@click.option("--offset", "offset", type=(float, float), multiple=True)
-	@click.option("--max-t", "maxt", type=int, default=10)
-	@click.option("--g90-extruder", "g90_extruder", is_flag=True)
-	@click.option("--progress", "progress", is_flag=True)
-	@click.argument("path", type=click.Path())
-	def main(path, speedx, speedy, speedz, offset, maxt, throttle, throttle_lines, g90_extruder, progress):
-		throttle_callback = None
-		if throttle:
-			def throttle_callback(filePos, readBytes):
-				if filePos % throttle_lines == 0:
-					# only apply throttle every $throttle_lines lines
-					time.sleep(throttle)
-
-		offsets = offset
-		if offsets is None:
-			offsets = []
-		elif isinstance(offset, tuple):
-			offsets = list(offsets)
-		offsets = [(0, 0)] + offsets
-		if len(offsets) < maxt:
-			offsets += [(0, 0)] * (maxt - len(offsets))
-
-		start_time = time.time()
-
-		progress_callback = None
-		if progress:
-			def progress_callback(percentage):
-				print("PROGRESS:{}".format(percentage))
-		interpreter = gcode(progress_callback=progress_callback)
-
-		interpreter.load(path,
-		                 speedx=speedx,
-		                 speedy=speedy,
-		                 offsets=offsets,
-		                 throttle=throttle_callback,
-		                 max_extruders=maxt,
-		                 g90_extruder=g90_extruder)
-
-		print("DONE:{}s".format(time.time() - start_time))
-		print("RESULTS:")
-		print(yaml.safe_dump(interpreter.get_result(), default_flow_style=False, indent="    ", allow_unicode=True))
-
-	# os args are gained differently on win32
-	try:
-		from click.utils import get_os_args
-		args = get_os_args()
-	except ImportError:
-		# for whatever reason we are running an older Click version?
-		args = sys.argv[1:]
-
-	if len(args) >= len(sys.argv):
-		# Now some ugly preprocessing of our arguments starts. We have a somewhat difficult situation on our hands
-		# here if we are running under Windows and want to be able to handle utf-8 command line parameters (think
-		# plugin parameters such as names or something, e.g. for the "dev plugin:new" command) while at the same
-		# time also supporting sys.argv rewriting for debuggers etc (e.g. PyCharm).
-		#
-		# So what we try to do here is solve this... Generally speaking, sys.argv and whatever Windows returns
-		# for its CommandLineToArgvW win32 function should have the same length. If it doesn't however and
-		# sys.argv is shorter than the win32 specific command line arguments, obviously stuff was cut off from
-		# sys.argv which also needs to be cut off of the win32 command line arguments.
-		#
-		# So this is what we do here.
-
-		# -1 because first entry is the script that was called
-		sys_args_length = len(sys.argv) - 1
-
-		# cut off stuff from the beginning
-		args = args[-1 * sys_args_length:] if sys_args_length else []
-
-	main(args=args, prog_name="octoprint-gcode-analysis")
